+++ kbl2vec/src/main/java/com/foursoft/harness/kbl2vec/transform/contacting/ContactingSpecificationTransformer.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,41 +23,38 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.harness.kbl2vec.transform.harness;
+package com.foursoft.harness.kbl2vec.transform.contacting;
 
-import com.foursoft.harness.kbl.v25.KblConnectorOccurrence;
 import com.foursoft.harness.kbl.v25.KblContactPoint;
 import com.foursoft.harness.kbl.v25.KblHarness;
-import com.foursoft.harness.kbl2vec.core.TestConversionOrchestrator;
+import com.foursoft.harness.kbl2vec.core.TransformationContext;
+import com.foursoft.harness.kbl2vec.core.TransformationResult;
+import com.foursoft.harness.kbl2vec.core.Transformer;
 import com.foursoft.harness.vec.v2x.VecContactPoint;
 import com.foursoft.harness.vec.v2x.VecContactingSpecification;
-import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import java.util.List;
 
-class ContactingSpecificationTransformerTest {
+import static com.foursoft.harness.kbl2vec.transform.Fragments.abbreviatedClassName;
 
-    @Test
-    void should_transformContactingSpecification() {
-        // Given
-        final ContactingSpecificationTransformer transformer = new ContactingSpecificationTransformer();
-        final TestConversionOrchestrator orchestrator = new TestConversionOrchestrator();
+public class ContactingSpecificationTransformer implements Transformer<KblHarness, VecContactingSpecification> {
 
-        final KblHarness source = new KblHarness();
+    @Override
+    public TransformationResult<VecContactingSpecification> transform(final TransformationContext context,
+                                                                      final KblHarness source) {
+        final VecContactingSpecification destination = new VecContactingSpecification();
+        destination.setIdentification(
+                abbreviatedClassName(destination.getClass()) + "-" + source.getPartNumber());
 
-        final KblConnectorOccurrence connectorOccurrence = new KblConnectorOccurrence();
-        final KblContactPoint contactPoint = new KblContactPoint();
-        connectorOccurrence.getContactPoints().add(contactPoint);
-        source.getConnectorOccurrences().add(connectorOccurrence);
+        return TransformationResult.from(destination)
+                .withDownstream(KblContactPoint.class, VecContactPoint.class, () -> getContactPoints(source),
+                                VecContactingSpecification::getContactPoints)
+                .build();
+    }
 
-        final VecContactPoint vecContactPoint = new VecContactPoint();
-        orchestrator.addMockMapping(contactPoint, vecContactPoint);
-
-        // When
-        final VecContactingSpecification result = orchestrator.transform(transformer, source);
-
-        // Then
-        assertThat(result).isNotNull()
-                .satisfies(v -> assertThat(v.getContactPoints()).containsExactly(vecContactPoint));
+    private List<KblContactPoint> getContactPoints(final KblHarness harness) {
+        return harness.getConnectorOccurrences().stream()
+                .flatMap(c -> c.getContactPoints().stream())
+                .toList();
     }
 }
