+++ kbl2vec/src/test/java/com/foursoft/harness/kbl2vec/transform/contacting/ContactPointTransformerTest.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.harness.kbl2vec.transform.connectivity;
+package com.foursoft.harness.kbl2vec.transform.contacting;
 
 import com.foursoft.harness.kbl.v25.KblContactPoint;
 import com.foursoft.harness.kbl.v25.KblProcessingInstruction;
